Move isRepresentable
